feat(fxWeaver): add javafx-weaver-spring-boot-starter package so beans in controllers are configured
`FXMLoader` was previously creating the controller but this was not 
managed by spring, so attempting to access services in controllers other 
than the welcome stage would result in a null pointer. using `FxWeaver` 
creates the controllers for us as it is a spring-managed bean and allow 
us to fully utilize spring. existing controllers have been updated to 
include `@FxmlView` and also pass the controller when using the 
navigation helper.