feat(images): include new images column in items table
items database will now contain the path to the items image, which will be used to replace the placeholder present when displaying an item in more detail. uploaded current set of images that is currently used by different items in the database. additionally, new inital data set has been dumped to include the new column.